+++ src/MetadataProcessor/ServiceCollectionExtensions.cs
@@ -1,3 +1,4 @@
+using Kurmann.Videoschnitt.MetadataProcessor.Services;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
